+++ questions/02-not-an-array-linked-list.js
@@ -1,0 +1,18 @@
+// Implement a class called NotAnArrayLinkedList...
+// All methods must be O(1)
+
+class NotAnArrayLinkedList {
+  constructor() {
+    this.storage = {};
+  }
+
+  push() {}
+
+  pop() {}
+
+  unshift() {}
+
+  shift() {}
+}
+
+module.exports = NotAnArrayLinkedList;
