+++ test/02-not-an-array-linked-list.test.js
@@ -1,0 +1,10 @@
+const assert = require('chai').assert;
+const NotAnArrayLinkedList = require('../questions/02-not-an-array-linked-list');
+
+describe('NotAnArrayLinkedList', () => {
+    describe('#Passing', () => {
+        it('should return true', () => {
+          return true;
+        });
+    });
+});
